requirements.txt
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.5
django-admirarchy==1.0.0
django-cors-middleware==1.3.1
django-filter==2.0.0
django-oauth-toolkit==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.9.1
djoser==1.4.0
future==0.17.1
idna==2.8
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10
MarkupSafe==1.1.0
mccabe==0.6.1
oauthlib==3.0.0
parso==0.3.1
pluggy==0.8.1
pycodestyle==2.4.0
pydocstyle==3.0.0
pyflakes==2.0.0
python-jsonrpc-server==0.1.2
python-language-server==0.21.5
pytz==2018.9
requests==2.21.0
rope==0.11.0
six==1.12.0
snowballstemmer==1.2.1
uritemplate==3.0.0
urllib3==1.24.1
yapf==0.25.0
